Many changes related to bookmarks
I created a proper handle for bookmakrs instead of using
optional<string> or whatever it was. I think all the bookmark
operations are working now.

I got a crash while saving, I think because pcloud was messing with
files being updated. I changed the logic to create a copy of the pdf,
open that one, and then write to the actual pdf. so far no problems.