Better handling of bookmarks process
logging error if we crash, ensure it shutsdown even if we crash.

Made running in visual studio work - it wasn't launching in the
bin folder, I don't quite understand  how it was running, but I changed
it to the bin dir, which then made me define QT_PLUGIN_PATH for it
to find the platform subdir. don't understand, but it works now.