Do not process page_loaded if it doesn't belong to you!
There may be better/more fixes, but while loading documents every
pdf viewer gets notified for every frigging page. Here we ignore
the call if not on a page we are currently displaying. I'd prefer
not to get the call at all, and to also ignore if tab is not current,
but maybe later.