Redraw app when computer wakes up
Trying a different way to notice when the computer wakes from sleep.
When it wakes the document is often rendered incorrectly, I guess
because the monitor gets switched to a differrent resolution while
sleeping? I don't know.

I removed a bunch of trace calls, they utterly dominate the logs
and added no value to understanding what is happening.