Tour supports svg animations
Note: there is a but in qt svg viewer (I believe). tabview.svg displays
correctly in all browsers, in VS code, etc. But incorrectly in
the qt web engine. I had to hand edit the svg: search for FFFF33 fill -
that was set to ffffff. I don't know, maybe it is a draw.io bug, in any
case this hand edit makes the svg display with the 'stick note' corrctly
yellow instead of white.

added a tour card to show toggling display of tabs.