Fixed shortcut for showing/hiding tabs
It was set twice, which seems to conflict. Did some code cleanup
because setting global shortcuts was strewn around a bit.

Now I add the shortcut to the menu, then call addAction(menu_shortcut)
which allows both the menu to display the shortcut key, and the action
to be global.