Fixed zooming on pages with blank space at bottom
The blank space at the bottom led it to conclude there was "too much"
too zoom and it wouldn't zoom the pages.

there may be cases where we need to not zoom, but I want to see them,
not make assumptions. Hence zooming always happens at the moment if
turned on.