Handle exceptions in muPDF bookmark code
A PDF had damaged bookmarks, and muPDf threw an exception.

This caused the app to stop loading the document, and the UI
endlessly showed loading.

I fixed that, but the initialize_document has a lot of early
returns that I need to investigate and fix - you should get an
error or something at least.