several small changes
fix the goto page shortcut (should be G).

Try to  shut down faster by notifying viewers that a shut down
is in progress so it can ignore signals that a page is loaded.

Diable save/set bookmarks from text file if the document doesn't
have bookmarks (the latter) or if the txt file doesn't exist (the
former)