Faster document refreshes
Every document that was opened received and processed refresh
calls when config is saved. this meant documents that aren't
currently visible rerender, make calls to the scrollbar, page
count ctrl, etc. It now just refreshes the tab in view, and then
sets the rest dirty. then when activated it will refresh iff set
to dirty.